restructured project layout and file imports
