+++ client/src/components/forms/index.js
@@ -1,0 +1,2 @@
+export * from "./RegisterForm/RegisterForm"
+export * from "./LoginForm/LoginForm"
