+++ client/src/components/pages/index.js
@@ -1,0 +1,2 @@
+export * from "./AuthPage/AuthPage"
+export * from "./LandingPage/LandingPage"
