+++ client/src/schemas/index.js
@@ -1,0 +1,2 @@
+export * from "./loginSchema"
+export * from "./registerSchema"
