+++ client/src/schemas/loginSchema.js
@@ -1,6 +1,6 @@
 import * as yup from "yup"
 
-const loginFormSchema = yup.object().shape({
+export const loginSchema = yup.object().shape({
   username: yup
     .string()
     .trim()
@@ -11,4 +11,4 @@ const loginFormSchema = yup.object().shape({
     .required("password is required")
 })
 
-export default loginFormSchema
+export default loginSchema
